modify shapedrawable(add gradient support)

+++ src/gui/drawables/paintdrawable.cc
@@ -8,18 +8,18 @@ PaintDrawable::PaintDrawable(int color){
 }
 
 void PaintDrawable::setCornerRadius(float radius){
-    std::vector<int>radii;
+    std::vector<float>radii;
     if(radius>0)
-        for(int i=0;i<8;i++)radii.push_back((int)radius);
+        for(int i=0;i<8;i++)radii.push_back(radius);
     setCornerRadii(radii);
 }
 
-void PaintDrawable::setCornerRadii(const std::vector<int>& radii){
+void PaintDrawable::setCornerRadii(const std::vector<float>& radii){
     if(radii.size()==0){
         if(getShape())setShape(nullptr);
     }else{
         Rect inset={0,0,0,0};
-        setShape(new RoundRectShape(radii,inset,std::vector<int>()));
+        setShape(new RoundRectShape(radii,inset,std::vector<float>()));
     }
 }
